Switch inventory logging to dedicated logger
The new format is much longer than before and is kind of spammy, so it
was a bit annoying in the console.

This switches it to a dedicated logger. With default configuration this
means the inventory logs won't show up at all, and with the right log4j
configuration it'll show up in its own files.